# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost:8443
    description: Generated server url
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication API
      summary: Refresh token
      description:
        Issues fresh credentials using a refresh token from cookie or authorization
        header.
      operationId: refresh
      parameters:
        - name: RefreshToken
          in: cookie
          required: false
          schema:
            type: string
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication API
      summary: Login
      description: Authenticates a user with username and password.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - SUCCESS
            - FAILURE
        message:
          type: string
        error:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...